CTCP-4758: Moving pre-cross-check questions to dedicated section in user answers.
